+++ src/main/java/si/matjazcerkvenik/alertmonitor/data/MemoryDataManager.java
@@ -1,3 +1,18 @@
+/*
+   Copyright 2021 Matjaž Cerkvenik
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
 package si.matjazcerkvenik.alertmonitor.data;
 
 import si.matjazcerkvenik.alertmonitor.model.DEvent;
@@ -12,12 +27,12 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class InMemoryDataManager implements IDataManager {
+public class MemoryDataManager implements IDataManager {
 
     private static SimpleLogger logger = LogFactory.getLogger();
 
-    public InMemoryDataManager() {
-        logger.info("InMemoryDataManager initialized");
+    public MemoryDataManager() {
+        logger.info("MemoryDataManager initialized");
     }
 
     /** List of webhook messages in its raw form. */
@@ -40,7 +55,7 @@ public class InMemoryDataManager implements IDataManager {
     public void addToJournal(List<DEvent> events) {
         journal.addAll(events);
         for (DEvent e : events) {
-            LogFactory.getLogger().info("InMemoryDataManager: adding to journal: uid=" + e.getUid());
+            LogFactory.getLogger().info("MemoryDataManager: adding to journal: uid=" + e.getUid());
         }
 
     }