+++ src/main/java/si/matjazcerkvenik/alertmonitor/model/config/ConfigReader.java
@@ -1,3 +1,18 @@
+/*
+   Copyright 2021 Matjaž Cerkvenik
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
 package si.matjazcerkvenik.alertmonitor.model.config;
 
 import org.yaml.snakeyaml.Yaml;
@@ -22,7 +37,7 @@ public class ConfigReader {
         try {
             inputStream = new FileInputStream(f);
             YamlConfig config = yaml.load(inputStream);
-            LogFactory.getLogger().info("providers config loaded: " + config.toString());
+            LogFactory.getLogger().info("providers config loaded: " + path);
             return config;
         } catch (FileNotFoundException e) {
             LogFactory.getLogger().warn("no providers.yml found at " +  path);