+++ src/main/java/si/matjazcerkvenik/alertmonitor/model/config/ProviderConfig.java
@@ -1,3 +1,18 @@
+/*
+   Copyright 2021 Matjaž Cerkvenik
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
 package si.matjazcerkvenik.alertmonitor.model.config;
 
 import java.util.Map;
@@ -6,7 +21,7 @@ public class ProviderConfig {
 
     private String id;
     private String name;
-    private String source;
+    private String type;
     private String uri;
     private Map<String, Object> params;
 
@@ -26,12 +41,12 @@ public class ProviderConfig {
         this.name = name;
     }
 
-    public String getSource() {
-        return source;
+    public String getType() {
+        return type;
     }
 
-    public void setSource(String source) {
-        this.source = source;
+    public void setType(String type) {
+        this.type = type;
     }
 
     public String getUri() {
@@ -55,7 +70,7 @@ public class ProviderConfig {
         return "ProviderConfig{" +
                 "id='" + id + '\'' +
                 ", name='" + name + '\'' +
-                ", source='" + source + '\'' +
+                ", type='" + type + '\'' +
                 ", uri='" + uri + '\'' +
                 ", params=" + params +
                 '}';